# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/user-clothing/save-interaction:
    post:
      tags:
      - User Clothing API
      summary: Save user interaction with clothing
      description: "Stores an interaction (e.g., LIKE, SKIP) between a user and a\
        \ clothing item. Returns HTTP 409 if interaction already exists."
      operationId: saveInteraction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserClothingDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/user-clothing/get-interactions:
    get:
      tags:
      - User Clothing API
      summary: Get user interactions by type
      description: "Retrieves a list of clothing items the user has interacted with,\
        \ filtered by interaction type."
      operationId: getInteractions
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: interactionType
        in: query
        required: true
        schema:
          type: string
          enum:
          - LIKED
          - DISLIKED
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/swipe-style/get-clothing-id:
    get:
      tags:
      - SwipeStyle API
      summary: Get clothing entity ID from productId
      description: Returns the internal database ID for a given productId.
      operationId: getClothingId
      parameters:
      - name: productId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
...
    UserClothingDTO:
      required:
      - clothingId
      - interactionType
      - userId
      type: object
      properties:
        userId:
          type: integer
          format: int64
        clothingId:
          type: integer
          format: int64
        interactionType:
          type: string
          enum:
          - LIKED
          - DISLIKED
...